# base-go/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {
            "name": "Direktorat Pengembangan Teknologi dan Sistem Informasi (DPTSI) - ITS",
            "url": "http://its.ac.id/dptsi",
            "email": "dptsi@its.ac.id"
        }
    },
    "paths": {
...
        "/spesimen": {
            "post": {
                "description": "Membuat spesimen tanda tangan baru untuk SDM",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Spesimen"
                ],
                "summary": "Membuat spesimen tanda tangan baru",
                "parameters": [
                    {
                        "description": "Data spesimen",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.CreateSpesimenRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.SpesimenTandaTangan"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/spesimen/sdm/{idSdm}": {
            "get": {
                "description": "Mendapatkan spesimen tanda tangan berdasarkan ID SDM",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Spesimen"
                ],
                "summary": "Mendapatkan spesimen tanda tangan berdasarkan ID SDM",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID SDM",
                        "name": "idSdm",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.SpesimenTandaTangan"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/spesimen/{id}": {
            "get": {
                "description": "Mendapatkan spesimen tanda tangan berdasarkan ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Spesimen"
                ],
                "summary": "Mendapatkan spesimen tanda tangan berdasarkan ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID Spesimen",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.SpesimenTandaTangan"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "put": {
                "description": "Memperbarui spesimen tanda tangan berdasarkan ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Spesimen"
                ],
                "summary": "Memperbarui spesimen tanda tangan",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID Spesimen",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Data update spesimen",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.UpdateSpesimenRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.SpesimenTandaTangan"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "description": "Menghapus spesimen tanda tangan berdasarkan ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Spesimen"
                ],
                "summary": "Menghapus spesimen tanda tangan",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID Spesimen",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "models.SpesimenTandaTangan": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "data": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "id_sdm": {
                    "type": "string"
                },
                "id_spesimen": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "updater": {
                    "type": "string"
                }
            }
        },
        "requests.CreateSpesimenRequest": {
            "type": "object",
            "required": [
                "data",
                "id_sdm"
            ],
            "properties": {
                "data": {
                    "type": "string"
                },
                "id_sdm": {
                    "type": "string"
                },
                "updater": {
                    "type": "string"
                }
            }
        },
        "requests.UpdateSpesimenRequest": {
            "type": "object",
            "required": [
                "data",
                "updater"
            ],
            "properties": {
                "data": {
                    "type": "string"
                },
                "updater": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "CSRF Token": {
            "description": "CSRF token yang didapatkan dari browser -\u003e inspect element -\u003e application -\u003e storage -\u003e cookies -\u003e CSRF-TOKEN (Untuk firefox, storage berada pada tab tersendiri)",
            "type": "apiKey",
            "name": "x-csrf-token",
            "in": "header"
        }
    },
...
}